Setup a basic view controller state
